reinstate Debian codename symlinks (closes: #1073785)
This reverts the removal of Debian codename symlinks from
commit 2d3eae916af51cc49ab0989cea1f5bfb58012179
because their absence breaks debootstrap when run in Debian-Installer
(which does not have `distro-info` available).

It seems entierly plausible that the absence of the Ubuntu symlinks will
similarly break debian-installer when used to install Ubuntu, but I'll
leave those as they are until that's confirmed by someone from Ubuntu.