missing ubuntu-distro-info fallback to endoflife.date
As the distro running debootstrap may not contain ubuntu-disto-info
use the information on endoflie.date for the ubuntu.

Ubuntu moves things to old-releases once it past the extended support
date, or if not, just the supported date.

If neither ubuntu-distro-info is there, or curl/jq, then assume
a supported distro on the assumption they are more used.