create /dev/fd symlink with --no-target-directory
Without the `-T` or `--no-target-directory` option, the last argument of
`ln` is not the link name but the directory in which to create the link.
Only if the given directory does not exist, does it become the link
name. For debootstrap this becomes a problem if `/dev/fd` already exists
in `TARGET` and if it does, it will usually be a symlink to
`/proc/self/fd`. So running

    ln -sf /proc/self/fd "$TARGET/dev/fd

Will attempt creating a link named `$TARGET/dev/fd/fd`. To avoid this,
run `ln` with `--no-target-directory`.