Self-identify in User-Agent sent by wget
Adding a custom user-agent will give the admin of debian (or its variant) mirrors a better understanding of the source of the requests. It is especially useful for mirror sites facing heavy traffic.

Comparing to apt-get, the behavior of debootstrap differs from apt in the following aspects, which makes the requests suspicious for admins without knowledge of debootstrap: 1. the number of consecutive requests made by debootstrap is usually larger than a single apt-get run, since bootstrapping a debian system would require more packages; 2. the user-agent header in the requests made by apt-get clearly states the tool initiating the requests, while debootstrap utilizes wget to send out the requests only containing Wget/<version>, which can be confusing with some abuse activities on mirror sites.